requirements.txt
# Core Flask dependencies
Flask==3.*
Flask-SQLAlchemy==3.*
Flask-CORS==4.*
Flask-APScheduler==1.*
python-dotenv==1.*

# WSGI server for Windows
waitress==3.*

# LlamaIndex and AI dependencies
llama-index-core==0.10.*
llama-index-llms-gemini==0.1.*
llama-index-embeddings-gemini==0.1.*
llama-index-readers-file==0.1.*
llama-index-vector-stores-chroma==0.1.*

# Vector database
chromadb>=0.4.15

# PDF processing
PyPDF2==3.*
pdfplumber==0.10.*
pdfminer.six

# File type detection for Windows
python-magic-bin==0.4.*

# Windows-specific utilities
pywin32==306

# Database
SQLAlchemy==2.*

# Utilities
requests==2.*
pathlib-abc==0.1.*

# Development and testing
pytest==7.*
pytest-asyncio==0.21.*
pytest-flask==1.*
black==23.*
flake8==6.*
vcrpy==5.*

# Optional: FAISS (install via conda if needed)
# faiss-cpu
